+++ packages/core/src/main/conversation/ConversationService.ts
@@ -118,11 +118,7 @@ export default class ConversationService {
       messageId,
     });
 
-    const preKeyBundles = await this.getPreKeyBundles(conversationId);
-    const plainTextBuffer: Buffer = this.protocolBuffers.GenericMessage.encode(ping).finish();
-    const payload: EncryptedAsset = await AssetCryptography.encryptAsset(plainTextBuffer);
-
-    await this.sendExternalGenericMessage(this.clientID, conversationId, payload, <UserPreKeyBundleMap>preKeyBundles);
+    await this.sendGenericMessage(this.clientID, conversationId, ping);
     return messageId;
   }
 
@@ -234,7 +230,8 @@ export default class ConversationService {
     return this.sendMessage(sendingClientId, conversationId, recipients);
   }
 
-  // TODO: The correct functionality of this function is heavily based on the case that it always runs into the catch block
+  // TODO: The correct functionality of this function is heavily based on the case that it always runs into the catch
+  // block
   private getPreKeyBundles(conversationId: string): Promise<ClientMismatch | UserPreKeyBundleMap> {
     return this.apiClient.conversation.api.postOTRMessage(this.clientID, conversationId).catch((error: AxiosError) => {
       if (error.response && error.response.status === 412) {