+++ packages/proteus/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="7.1.52"></a>
+## [7.1.52](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@7.1.51...@wireapp/proteus@7.1.52) (2018-06-05)
+
+
+
+
+**Note:** Version bump only for package @wireapp/proteus
+
+<a name="7.1.51"></a>
+## [7.1.51](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@7.1.50...@wireapp/proteus@7.1.51) (2018-06-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/proteus
+
+<a name="7.1.50"></a>
+## [7.1.50](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@7.1.49...@wireapp/proteus@7.1.50) (2018-06-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/proteus
+
+<a name="7.1.49"></a>
+## [7.1.49](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@7.1.48...@wireapp/proteus@7.1.49) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @wireapp/proteus
+
 <a name="7.1.48"></a>
 ## [7.1.48](https://github.com/wireapp/wire-web-packages/tree/master/packages/proteus/compare/@wireapp/proteus@7.1.47...@wireapp/proteus@7.1.48) (2018-05-30)