+++ packages/travis-bot/CHANGELOG.md
@@ -3,6 +3,81 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.5"></a>
+## [1.0.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@1.0.4...@wireapp/travis-bot@1.0.5) (2018-06-05)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
+<a name="1.0.4"></a>
+## [1.0.4](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@1.0.3...@wireapp/travis-bot@1.0.4) (2018-06-05)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
+<a name="1.0.3"></a>
+## [1.0.3](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@1.0.2...@wireapp/travis-bot@1.0.3) (2018-06-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
+<a name="1.0.2"></a>
+## [1.0.2](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@1.0.1...@wireapp/travis-bot@1.0.2) (2018-06-04)
+
+
+### Bug Fixes
+
+* **travis-bot:** ClientType import ([#662](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/issues/662)) ([1ec51bf](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/commit/1ec51bf))
+
+
+
+
+<a name="1.0.1"></a>
+## [1.0.1](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@1.0.0...@wireapp/travis-bot@1.0.1) (2018-06-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
+<a name="1.0.0"></a>
+# [1.0.0](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@0.1.136...@wireapp/travis-bot@1.0.0) (2018-06-04)
+
+
+### Code Refactoring
+
+* **api-client,core:** Add ClientType.NONE ([#660](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/issues/660)) ([97e7b12](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/commit/97e7b12))
+
+
+### BREAKING CHANGES
+
+* **api-client,core:** Expecting ClientType in init calls, register calls and LoginData type.
+
+
+
+
+<a name="0.1.136"></a>
+## [0.1.136](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@0.1.135...@wireapp/travis-bot@0.1.136) (2018-06-04)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
+<a name="0.1.135"></a>
+## [0.1.135](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@0.1.134...@wireapp/travis-bot@0.1.135) (2018-06-01)
+
+
+
+
+**Note:** Version bump only for package @wireapp/travis-bot
+
 <a name="0.1.134"></a>
 ## [0.1.134](https://github.com/wireapp/wire-web-packages/tree/master/packages/travis-bot/compare/@wireapp/travis-bot@0.1.133...@wireapp/travis-bot@0.1.134) (2018-05-30)